+++ ent/loginsession.go
@@ -17,11 +17,11 @@ import (
 type LoginSession struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	ID int64 `json:"id,omitempty"`
 	// CreateTime holds the value of the "create_time" field.
 	CreateTime time.Time `json:"create_time,omitempty"`
 	// UserID holds the value of the "user_id" field.
-	UserID int `json:"user_id,omitempty"`
+	UserID int64 `json:"user_id,omitempty"`
 	// Edges holds the relations/edges for other nodes in the graph.
 	// The values are being populated by the LoginSessionQuery when eager-loading is set.
 	Edges        LoginSessionEdges `json:"edges"`
@@ -79,7 +79,7 @@ func (ls *LoginSession) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			ls.ID = int(value.Int64)
+			ls.ID = int64(value.Int64)
 		case loginsession.FieldCreateTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field create_time", values[i])
@@ -90,7 +90,7 @@ func (ls *LoginSession) assignValues(columns []string, values []any) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field user_id", values[i])
 			} else if value.Valid {
-				ls.UserID = int(value.Int64)
+				ls.UserID = value.Int64
 			}
 		default:
 			ls.selectValues.Set(columns[i], values[i])
