+++ ent/loginsession/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.LoginSession {
+func ID(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.LoginSession {
+func IDEQ(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.LoginSession {
+func IDNEQ(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.LoginSession {
+func IDIn(ids ...int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.LoginSession {
+func IDNotIn(ids ...int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.LoginSession {
+func IDGT(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.LoginSession {
+func IDGTE(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.LoginSession {
+func IDLT(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.LoginSession {
+func IDLTE(id int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldLTE(FieldID, id))
 }
 
@@ -61,7 +61,7 @@ func CreateTime(v time.Time) predicate.LoginSession {
 }
 
 // UserID applies equality check predicate on the "user_id" field. It's identical to UserIDEQ.
-func UserID(v int) predicate.LoginSession {
+func UserID(v int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldEQ(FieldUserID, v))
 }
 
@@ -106,22 +106,22 @@ func CreateTimeLTE(v time.Time) predicate.LoginSession {
 }
 
 // UserIDEQ applies the EQ predicate on the "user_id" field.
-func UserIDEQ(v int) predicate.LoginSession {
+func UserIDEQ(v int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldEQ(FieldUserID, v))
 }
 
 // UserIDNEQ applies the NEQ predicate on the "user_id" field.
-func UserIDNEQ(v int) predicate.LoginSession {
+func UserIDNEQ(v int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldNEQ(FieldUserID, v))
 }
 
 // UserIDIn applies the In predicate on the "user_id" field.
-func UserIDIn(vs ...int) predicate.LoginSession {
+func UserIDIn(vs ...int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldIn(FieldUserID, vs...))
 }
 
 // UserIDNotIn applies the NotIn predicate on the "user_id" field.
-func UserIDNotIn(vs ...int) predicate.LoginSession {
+func UserIDNotIn(vs ...int64) predicate.LoginSession {
 	return predicate.LoginSession(sql.FieldNotIn(FieldUserID, vs...))
 }
