+++ ent/problemcase/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.ProblemCase {
+func ID(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.ProblemCase {
+func IDEQ(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.ProblemCase {
+func IDNEQ(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.ProblemCase {
+func IDIn(ids ...int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.ProblemCase {
+func IDNotIn(ids ...int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.ProblemCase {
+func IDGT(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.ProblemCase {
+func IDGTE(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.ProblemCase {
+func IDLT(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.ProblemCase {
+func IDLTE(id int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldLTE(FieldID, id))
 }
 
@@ -75,7 +75,7 @@ func IsDeleted(v bool) predicate.ProblemCase {
 }
 
 // ProblemID applies equality check predicate on the "problem_id" field. It's identical to ProblemIDEQ.
-func ProblemID(v int) predicate.ProblemCase {
+func ProblemID(v int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldEQ(FieldProblemID, v))
 }
 
@@ -180,22 +180,22 @@ func IsDeletedNEQ(v bool) predicate.ProblemCase {
 }
 
 // ProblemIDEQ applies the EQ predicate on the "problem_id" field.
-func ProblemIDEQ(v int) predicate.ProblemCase {
+func ProblemIDEQ(v int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldEQ(FieldProblemID, v))
 }
 
 // ProblemIDNEQ applies the NEQ predicate on the "problem_id" field.
-func ProblemIDNEQ(v int) predicate.ProblemCase {
+func ProblemIDNEQ(v int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldNEQ(FieldProblemID, v))
 }
 
 // ProblemIDIn applies the In predicate on the "problem_id" field.
-func ProblemIDIn(vs ...int) predicate.ProblemCase {
+func ProblemIDIn(vs ...int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldIn(FieldProblemID, vs...))
 }
 
 // ProblemIDNotIn applies the NotIn predicate on the "problem_id" field.
-func ProblemIDNotIn(vs ...int) predicate.ProblemCase {
+func ProblemIDNotIn(vs ...int64) predicate.ProblemCase {
 	return predicate.ProblemCase(sql.FieldNotIn(FieldProblemID, vs...))
 }
