+++ ent/schema/group.go
@@ -14,9 +14,9 @@ type Group struct {
 // Fields of the Group.
 func (Group) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("id").Unique(),
+		field.Int64("id").Unique(),
 		field.String("group_name").Default("unknown"),
-		field.Int("root_group_id").Default(1),
+		field.Int64("root_group_id").Default(1),
 	}
 }
