+++ ent/submit/submit.go
@@ -92,15 +92,15 @@ func ValidColumn(column string) bool {
 
 var (
 	// StatusValidator is a validator for the "status" field. It is called by the builders before save.
-	StatusValidator func(int8) error
+	StatusValidator func(int16) error
 	// PointValidator is a validator for the "point" field. It is called by the builders before save.
 	PointValidator func(int16) error
 	// DefaultCreateTime holds the default value on creation for the "create_time" field.
 	DefaultCreateTime time.Time
 	// TotalTimeValidator is a validator for the "total_time" field. It is called by the builders before save.
-	TotalTimeValidator func(int) error
+	TotalTimeValidator func(int32) error
 	// MaxMemoryValidator is a validator for the "max_memory" field. It is called by the builders before save.
-	MaxMemoryValidator func(int) error
+	MaxMemoryValidator func(int32) error
 	// CaseVersionValidator is a validator for the "case_version" field. It is called by the builders before save.
 	CaseVersionValidator func(int8) error
 )
